+++ src/main/java/com/mastercard/developer/service/domain/GenericOffersCriterion.java
@@ -3,6 +3,8 @@ package com.mastercard.developer.service.domain;
 import lombok.Builder;
 import lombok.Data;
 
+import java.util.List;
+
 @Data
 @Builder
 public class GenericOffersCriterion {
@@ -15,4 +17,16 @@ public class GenericOffersCriterion {
   Integer itemsPerPage;
   Integer pageNumber;
   String lang;
+  Integer issuerIca;
+  String bankProductCode;
+  String offerType;
+  String category;
+  String offerCountry;
+  List<String> languages;
+  Integer offset;
+  Integer limit;
+  String sort;
+  String startDate;
+  String endDate;
+  String dateFilter;
 }